Update Arduino Driver
Changed the print statements to make everything print on the same line with the last print statement using println to add a new line character. This enables the python code to terminate its loop that reads the data in. I also moved the if else statement for the float sensor down from the top.